requirements.txt
numpy==1.23.3
opencv-python==4.6.0.66
playsound
gTTS
pillow